+++ development.ini
@@ -14,8 +14,8 @@ pyramid.default_locale_name = en
 pyramid.includes =
     pyramid_debugtoolbar
 
-# sqlalchemy.url = sqlite:///%(here)s/translateApp.sqlite
-sqlalchemy.url = postgresql+pypostgresql://pyramid_user:password@localhost/project_db
+sqlalchemy.url = sqlite:///%(here)s/translateApp.sqlite
+#sqlalchemy.url = postgresql+pypostgresql://pyramid_user:password@localhost/project_db
 
 auth.secret = seekrit
