+++ translateapp/models/__init__.py
@@ -5,7 +5,8 @@ import zope.sqlalchemy
 
 # import or define all models here to ensure they are attached to the
 # Base.metadata prior to any initialization routines
-from .models import Users  # noqa
+from .user import User
+from .phrase import Phrase
 
 # run configure_mappers after defining all of the models to ensure
 # all relationships can be setup
