+++ translateapp/models/user.py
@@ -1,0 +1,11 @@
+from sqlalchemy import Column, Integer, Text
+
+from .meta import Base
+
+
+class User(Base):
+    __tablename__ = 'users'
+    id = Column(Integer, primary_key=True)
+    name = Column(Text, nullable=False)
+    email = Column(Text, nullable=False)
+    password = Column(Text, nullable=False)
