+++ translateapp/scripts/initializedb.py
@@ -15,7 +15,8 @@ from ..models import (
     get_session_factory,
     get_tm_session,
     )
-from ..models import models
+
+from ..models import User
 
 
 def usage(argv):
@@ -41,5 +42,5 @@ def main(argv=sys.argv):
     with transaction.manager:
         dbsession = get_tm_session(session_factory, transaction.manager)
 
-        model = MyModel(name='one', value=1)
+        model = User(name='mijin', email='mijin@mijin.com', password='mijin')
         dbsession.add(model)
